Add new records 2024 09

+++ lib/tasks/add_records.rake
@@ -1,19 +1,30 @@
 namespace :lester do
-  desc 'Destroy All Records'
-  task destroy_all: :environment do
-    Transfer.destroy_all
-    Membership.destroy_all
-    Group.destroy_all
-    Person.destroy_all
+  desc "Augment"
+  task augment: :environment do
+    # From a file created and maintained by me, whenever noteworthy people or groups appear in the news
+    FileIngestor.general_upload('csv_data/new_people_groups_positions-2024-09-13.csv')
 
-    ActiveRecord::Base.connection.reset_pk_sequence!('transfers')
-    ActiveRecord::Base.connection.reset_pk_sequence!('memberships')
-    ActiveRecord::Base.connection.reset_pk_sequence!('groups')
-    ActiveRecord::Base.connection.reset_pk_sequence!('people')
+
+    # From a file created and maintained by me, linking groups to groups, whenever information comes to light
+    FileIngestor.affiliations_upload('csv_data/affiliations-2024-09-18.csv')
+
+
+    # This is a reminder, to add more info on ministries, should also add what the ministry is responsible for (also existing records)
+    federal_ministries = [
+      # 'csv_data/ministries_morrison.csv',
+      # 'csv_data/ministries_turnbull.csv',
+    ]
+
+    # using data sourced from wikipedia
+    federal_ministries.each do |file|
+      FileIngestor.ministries_upload(file)
+    end
   end
 
   desc "Destroy all records and re-populate"
   task populate: :environment do
+
+    raise if Rails.env.production?
 
     Transfer.destroy_all
     Position.destroy_all
